packages/tools/skill-2-data-transformer/requirements.txt
openai
python-dotenv
asyncio
